+++ app/src/main/java/com/leoevg/geoquiz/domain/repository/UserRepository.kt
@@ -4,7 +4,7 @@ import com.google.firebase.auth.AuthResult
 
 interface UserRepository {
     // ф-я suspend - т.к. была асинхронная таска в FIREBASE
-    suspend fun getMaxResultByUserId(email: String): Double?
-    suspend fun updateMaxResultByUserId(email: String, newMaxScore: Double): Boolean
+    suspend fun getMaxResultByUserId(email: String): Int?
+    suspend fun updateMaxResultByUserId(email: String, newMaxScore: Int): Boolean
 }
