+++ app/src/main/java/com/leoevg/geoquiz/navigation/NavigationPaths.kt
@@ -10,7 +10,7 @@ sealed interface NavigationPaths{
 
       @Serializable data class Quiz(val typeGame: TypeGame): NavigationPaths
 
-      @Serializable data class Finish(val finalScore: Double): NavigationPaths
+      @Serializable data class Finish(val finalScore: Int): NavigationPaths
       @Serializable data object Rate: NavigationPaths
       @Serializable data object Userpic: NavigationPaths
 }
