+++ app/src/main/java/com/leoevg/geoquiz/screens/question/QuestionScreenState.kt
@@ -12,7 +12,7 @@ data class QuestionScreenState (
     val isAnswerRight: Boolean? = null, // если мы еще не ответили - состояние null,
     // дальше смена на false/true
     val isHintUsed: Boolean = false,
-    val currentScore: Double = 0.0
+    val currentScore: Int = 0
 
 )
